requirements.txt
Faker==13.3.4
more-itertools==8.13.0
mysql-connector-python==8.0.29
-e git+https://github.com/eidos06/pglast.git@ac9cf0ce00b4ff30edc64242634700661faaf9e7#egg=pglast
protobuf==4.21.5
psycopg==3.0.11
psycopg2-binary==2.9.3
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0